backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
httpx==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
azure-storage-blob==12.23.0
azure-cosmos==4.7.0
azure-identity==1.18.0
Pillow==10.4.0
langgraph
neo4j
langchain-neo4j
langchain-google-genai
